+++ changelog.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Translation
 
+### Changed
+- Changed padding in pages
+
 
 
 ## [v0.1.3] - 29/5/2022